test(compiler-plugin/utils): move PreviewHintTestSupport to utils/ + update consumer imports
Ticket 1 で確定した production の `utils/` 集約方針に sync する。`PreviewHintTestSupport`
は output directory 走査 helper の集合 (`File.previewHint{Marker,Facade}{Files,Names}()`)
で、feature 横断で共有する test util なので `utils/` 配下が適切。

- old: `compiler-plugin/src/test/.../feature/PreviewHintTestSupport.kt`
       (package `feature`)
- new: `compiler-plugin/src/test/.../utils/PreviewHintTestSupport.kt`
       (package `utils`)

これに伴い 6 ファイルの import を新パスに追従更新:

- `feature/previewCollection/PreviewHintEnabledFlagTest.kt`
- `feature/previewCollection/PreviewHintScopeTest.kt`
- `feature/previewCollection/fir/hintAndMarkerGeneration/PreviewHintMarkerSealOrHiddenTest.kt`
- `feature/previewCollection/fir/hintGeneration/PreviewHintHiddenDeprecationTest.kt`
- `feature/previewCollection/fir/scopeValidation/DefaultCollectScopeDslTest.kt`
- `feature/previewCollection/ir/collectPreviewsReplacement/PreviewHintIgnoreCrossModuleTest.kt`

`previewHint*Files()` / `previewHint*Names()` は internal extension のままで、
同 source set 内 (`compiler-plugin/src/test/kotlin/`) からの参照は modifier の
変更不要。

検証:
- `:compiler-plugin:ktlintFormat` + `:compiler-plugin:ktlintCheck` pass
- `:compiler-plugin:test --rerun-tasks` pass (131 tests total)
- `apiCheck` pass
- `find compiler-plugin -name build/api -type d` で `compat-k*/build/api/` 以外ゼロ
- wildcard import 増加なし
- `grep -rln "compiler.feature.previewHint" compiler-plugin/src/test/kotlin/`
  で旧 import の残存ゼロ
