+++ compiler-plugin/src/test/kotlin/me/tbsten/compose/preview/lab/compiler/feature/previewCollection/fir/hintGeneration/PreviewHintHiddenDeprecationTest.kt
@@ -7,8 +7,8 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.shouldBe
 import me.tbsten.compose.preview.lab.compiler.CompilerPluginTestBase
-import me.tbsten.compose.preview.lab.compiler.feature.previewHintFacadeNames
-import me.tbsten.compose.preview.lab.compiler.feature.previewHintMarkerNames
+import me.tbsten.compose.preview.lab.compiler.utils.previewHintFacadeNames
+import me.tbsten.compose.preview.lab.compiler.utils.previewHintMarkerNames
 import me.tbsten.compose.preview.lab.compiler.isAtLeast
 
 /**
