+++ compiler-plugin/src/test/kotlin/me/tbsten/compose/preview/lab/compiler/feature/previewCollection/ir/collectPreviewsReplacement/PreviewHintIgnoreCrossModuleTest.kt
@@ -8,8 +8,8 @@ import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
 import io.kotest.matchers.shouldBe
 import me.tbsten.compose.preview.lab.compiler.CompilerPluginTestBase
-import me.tbsten.compose.preview.lab.compiler.feature.previewHintFacadeFiles
-import me.tbsten.compose.preview.lab.compiler.feature.previewHintMarkerFiles
+import me.tbsten.compose.preview.lab.compiler.utils.previewHintFacadeFiles
+import me.tbsten.compose.preview.lab.compiler.utils.previewHintMarkerFiles
 import me.tbsten.compose.preview.lab.compiler.isAtLeast
 import me.tbsten.compose.preview.lab.compiler.loadCollectedPreviews
